pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ietf-dtn-demo-agent"
version = "0.0.0"
authors = [
  { name="Brian Sipos", email="brian.sipos+ietf@gmail.com" },
]
description = "A demonstration agent for the DTN BPv7/TCPCLv4/UDPCLv2"
readme = "README.md"
license-files = ["LICENSE"]
requires-python = ">=3.7"
classifiers = [
  "Development Status :: 4 - Beta",
  "Intended Audience :: Developers",
  "Programming Language :: Python :: 3",
  "Operating System :: OS Independent",
  "Topic :: Software Development :: Libraries",
]

dependencies = [
  "asn1",
  "cbor2 <5.8", # work around cbor2 issue 272
  "pycose @git+https://github.com/BrianSipos/pycose.git@sipos-combined-patches",
  "crcmod ~=1.7",
  "cryptography >=41",
  "certvalidator",
  "dbus-python",
  "ifaddr ~=0.2",
  "ipaddress ~=1.0",
  "macaddress ~=2.0",
  # work around lack of release version fix for oscrypto issue #75
  # https://github.com/wbond/oscrypto/pull/76
  "oscrypto @git+https://github.com/wbond/oscrypto.git@master",
  "portion ~=2.1",
  "psutil",
  "pycryptodome ~=3.21",
  "PyGObject ~=3.34", # glib integration
  "PyYAML",
  "python3-dtls",
  "scapy ~=2.6",
  "six",
  "zeroconf ~=0.148",
]

[project.optional-dependencies]
test = [
  "pytest",
  "pytest-subtests",
  "pytest-cov",
  "coverage",
]
pylint = [
  "pylint",
  "pylint_junit",
]
format = [
  "autopep8",
  "licenseheaders",
]
flake8 = [
  "flake8",
  "flake8_pyproject",
]
docs = [
  "sphinx ~=8.2",
  "sphinx_rtd_theme",
  "sphinx-argparse",
  "sphinx-autodoc-typehints",
]
container = [
  "PyYAML",
  "asn1",
  "cryptography ~=46.0",
]

[tool.pytest.ini_options]
addopts = [
    "--import-mode=importlib",
]
log_level="debug"
log_cli="True"
log_cli_level="warning"

[tool.pylint.MAIN]
load-plugins="pylint_junit"
suggestion-mode="no"
[tool.pylint.FORMAT]
max-line-length=100
[tool.pylint.DESIGN]
min-public-methods=1

[tool.autopep8]
max_line_length=160

[tool.flake8]
max-line-length=160

[project.scripts]
tcpcl-messagegen = "tcpcl.test.messagegen:main"
tcpcl-bundlegen = "tcpcl.test.bundlegen:main"
tcpcl-agent = "tcpcl.cmd:main"
udpcl-agent = "udpcl.cmd:main"
btpu-agent = "btpu.cmd:main"
bp-agent = "bp.cmd:main"

[project.urls]
"Homepage" = "https://github.com/BrianSipos/dtn-demo-agent"
"Bug Tracker" = "https://github.com/BrianSipos/dtn-demo-agent/issues"
